pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.117.1",
    "google-genai>=1.39.1",
    "jinja2>=3.1.6",
    "langgraph>=0.6.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tavily-python>=0.3.3",
    "tenacity>=8.2.0",
    "uvicorn>=0.37.0",
]
